Added test for quasinormal step Cauchy safeguard
I added a test for when the quasinormal step has its Cauchy point
safeguarded.

In addition, I needed to rework GMRES to try and detect when something
goes wrong, which intersects with the Newton failed test.  Before, I had
the routine quit when the Gram-Schmidt coefficients became small
relative to the first.  Unfortunately, I have some other tests, where
these coefficients can differ by more than machine precision and this is
important.  That's, well, bad, so I modified things to just look for
NaNs instead.

Also, it appeared as though the stopping condition for the quasinormal
method was incorrect when we need safeguarding after hitting the
trust-region.  Here, we were not denoting that the safeguard was active.
Rather, we used the TR stopping condition and displayed an alpha < 1.  I
can't remember if I did that because I really wanted the trust-region to
open up even though we hit the safeguard or if this was a bug.  Anyway,
if we hit the trust-region and also have to cut due to the safeguard, we
now say that we hit the safeguard.