Added alpha_x_qn check to all quasinormal tests
I added a check on alpha_x_qn to all quasinormal tests.  We already had
a flag to see if we truncated things, but this new test makes sure that
we don't indicate that we truncated things when we don't.  This actually
found a small bug where this quantity was misreported when we hit the
LocalMin and Feasible stopping conditions.