Added better quasinormal safeguard tests
I added a better quasinormal safeguard test that checks whether or not
we've obeyed the fraction to the boundary rule.  In the cases where we
don't check the safeguard, we also make sure that we're not hitting the
boundary.

On top of this, I added a test where the quasinormal Newton step hits
the trust-region, but we still use the safeguarded step.