Fixed bug where qn reported dogleg when Cauchy
I fixed a bug where the quasinormal step reported a safeguarded dogleg
step when in fact it was a safeguarded Cauchy point.  This occured when
with hit the trust-region on the dogleg step, but both the dogleg step
and the Cauchy point were infeasible.  Here, we need to retreat back and
do a safeguard from the origin to the Cauchy point, which means that we
needed to denote that this was in fact the Cauchy point safeguarded.

In order to detect these bugs in the future, I added a new set of tests
that verifies whether or not we compute any augmented system solves.
That let's us make sure that we hit certain pieces of the code.

On top of all of this, I added a test for the above where we hit the
trust-region on the dogleg step, but both the dogleg and the Cauchy
point are infeasible, so we retreat to the Cauchy point.