Fixed QN bug where Newton alternative not computed
I fixed a bug where the alternative Newton step was not computed.
Basically, if we solve for the Newton solve, but are unable to use it,
we test to see if cutting back the Newton point to the origin gives a
better solution that what we computed with the dogleg method.
Previously, if we had to cut all of the way back to the Cauchy point
because of the safeguarding, we didn't compute this step.  Now, we do.

In order to test this, I added a test where we use the Newton point
interesected with the trust-region, but where we cut off the Cauchy
point with the inequality constraint.