Modified nullspace projection for TCG
I modified the nullspace projection for TCG to account for a couple of
different use cases

1.  When dx_t_uncorrected is already in the nullspace, we no longer do
    an augmented system solve

2.  When dx_t_uncorrected turns out to be zero, we don't do a solve
    either.  This can be hard to handle with 1 in the same way due to
    numerical errors in detecting 0.

3.  Modified how the algorithm reacts when the projected point turns out
    to be zero.  We now refer to both norm_dxtyp and eps_dx to determine
    when the step is small.  Note, this means that eps_dx affects this
    tolerance and should be set appropriately.

In addition, I wrote some code to soften the augmented system solve
stopping conditions.  GMRES really needs a stopping condition of the
form

    sqrt(|| e1 ||^2 + || e2 ||^2) <= xi * stuff.

However, the theory has a number of conditions of the form

    || e1 || + || e2 || <= xi * stuff.

Note, the lack of squares.  Now, technically, we don't really violate
anything if we use the bottom tolerance when we need the top condition
since

   || e1 ||^2 + || e2 ||^2
   <= || e1 ||^2 + || e2 ||^2 + 2 || e1 || || e2 ||
   = (|| e1 || + || e2 ||)^2

However, we're really forcing things to be smaller than they need to be.
It's possible to back walk the precise tolerance using the parallelogram
law, which is what the new code does.  I've not enabled it everywhere
yet, just the nullspace projection, while I implement these tests.

Finally, I started adding some tests for the nullspace projection.
Right now, some of it overlaps with the quasinormal test code, so I need
to clean some things up.