Added two more nullspace projection tests
I added two more nullspace projection tests.  The first insures that if
we already have a direction in the nullspace, that we don't do an
augmented system solve.  The second insures that zero directions are not
reprojected.  In theory, the first test should cover the second, but we
actually have separate code since detecting zeros can be numerically
difficult.

In addition, I removed the code that short circuits the quasinormal step
augmented system solve when the Cauchy point brought us to optimality.
With the updated quasinormal code, that should never happen because we
trap that case directly.