Combined projected gradient and proj for CG
I combined much of the code for projecting the gradient into the
nullspace of g'(x) as well as the nullspace projection for truncated CG.
In addition, I doubled the number of tests that we run for the nullspace
projection to run tests on both configurations of code.

In theory, the tangential step is another nullspace projection, but the
stopping conditions and requirements on this step are pretty different
from the above two.  As such, I'll need to assess later whether or not
that code should be folded in or not.