Fixed bug with deserialization in MATLAB
I fixed a bug with the deserialization in MATLAB.  In order to make the
restart file, for Rm vectors, match Python and C++, we should be
outputing a row vector.  Then, when we input, we should be receiving a
row vector.  Sometimes, when reading in, we'd parse out a row vector and
sometimes it'd be a column vector.  Honestly, I think this has something
to do with how long the vector was and whether or not we had newline
characters.  Anyway, that shouldn't be that way, so we now force the
vector to be a column vector when reading in and force it to be a row
vector when writing out.