First draft of MATLAB/Octave separation
This is sort of an ugly check-in, but it does separate the MATLAB and
Octave interfaes, so that they now compile and install separately.  In
addition, the tests now run both.  Now, that being said, MATLAB
occasionally throws the error:

dlopen: cannot load any more object with static TLS

when calling mex files.  This is a bug on Mathworks side.  The only easy
workaround that I know is to load Optizelle earlier in the life of a
MATLAB session.