Added a unit test for the build configuration
I added a unit test for the build configuration that we use in the
manual.  Then, I wrote a script to process this same configuration for
the manual directly.  On that topic, we now run the Rosenbrock example
and take the output directly for the manual.  This does require that the
C++ library be compiled, whereas we could just compile the manual
direclty before, but hopefully it's nicer since it forces things to be
in sync.