Added new preprocessing setup
I added a new preprocessing setup that'll hopefully make it easier to
keep certain pieces of the documentation and code consistent.
Basically, it's some scripts and code that preprocess source files.  At
the end of the day, there's a C++ source file that handles the
preprocessing, so we should be able to process arbitrary information
whether it comes from json, database, or text files.  For the immediate
check-in, I'm preprocessing some files based on text information.
Basically, it's a better way to keep the configure unit test and the
documentation consistent.