Fixed assortment of bugs found on Windows
I fixed an assortment of bugs that arose from compiling and running
everything on Windows.  Also, notably, I fixed the scripts so that our
installer for Windows should work correctly with the 64-bit versions of
Octave for Windows floating around.  Beyond that, it looks like some
files weren't being copied over for the unit tests, so hoepfully those
bugs are fixed now.  I also fixed some of the error checking with file
IO.  We're somewhat overdue for some code refurbishing that fixes all of
our IO and error checking elsewhere in the code.