Updated documentation with complete license info
I updated the documentation to include the complete license info for all
of the dependencies.  This surplanted much of the stuff in the section
on dependencies.  Technically, we can change out things like BLAS/LAPACK
or the C++ compilers and have a different set of licenses.  Before, I
had links to other packages that were compatible.  However, really, most
people aren't going to do that and our premade installers are going to
use what we have here.  As such, it seems to be somewhat more useful for
people to see the most common use cases, and what we have in our
installer, along with the license information for the complete library.
At least, that's what I get asked by legal when I work with companies.