Reworked C++ and Python error handling
I reworked the C++ and Python error handling.  At it's core, we:

1.  Changed the messaging object to be a function that accepts a string
    and then outputs it to somewhere

2.  Since the Messaging object no longer has an error function, we now
    throw exceptions when there's an error

Part of the tricky part of that is that there are a ton of functions
that accepted a messaging object for the error handling.  All of that
has disappeared.  In addition, the Python interface was swallowing
Python exceptions thrown by the user code, which was difficult on the
users since it made diagnosting errors difficult.  Now, we throw two
different kinds of exceptions in the Python code to distiguish between
user exceptions and Optizelle exceptions.

This change, by the way, will break everyone's code.  I'm really sorry
about that, but I do believe that the change is for the better.