Reworked MATLAB/Octave error handling and memory
I revised the MATLAB/Octave error handling to always throw the native
exception mechanism (error) rather than using the Messaging object.
This simplifies the messaging object to just be a function and also puts
it in line with the Python and C++ versions.

In addition, I modified how memory is managed on the MATLAB/Octave mex
code.  Basically, MATLAB/Octave manage memory very similar to C.
There's an allocation function and then there's a deallocation function.
The trick is that some things should be deallocated, but others should
not.  In general, things like looking into a structure array should not
be deallocated, but creating a matrix should.  Anyway, before, there was
this concept of attached or captured.  Now, it's unmanaged and managed,
respectively.  For the most part, I have everything ported over to use
this scheme rather than using mxArray * directly.  That being said,
often I need to put these new classes inside of a shared or unique C++
pointer.  That works fine except that mxArray * is really just a typedef
for void *.  As such, if I dereference the shared or unique pointer
rather than our managed and unmanaged pointers, there's not a compiler
error and things just crash.  Hopefully, all of those bugs have been
eliminated, but that's something that's unfortunate and could be an
ongoing issue.  Maybe I need a dereferencing operator not called get to
catch the error.

Outside of that, I should have updated all of the unit tests and
examples to use the new scheme as well.