Overhauled MATLAB/Octave memory management (again)
I did another overhaul of the MATLAB/Octave memory management to
hopefully simplify things and catch more errors.  Before, I ended up
splitting our pointer type into managed and unmanaged pieces.  This
worked, but it had some really awkward constructors that I wasn't happy
about.  Now, I put the mxArray pointer into a structure along with
whatever destructor we're using.  This is nothing for unmanaged memory
and mxDestroyArray for managed memory.  Then, I put this structure into
a shared pointer which is held in our mxArrayPtr type.  In this way, we
can basically do safe soft copies.  On top of this, like Python, I
isolated all of the mex routines that we need and put them into the capi
namespace.  Hopefully, I'm trapping all of the appropriate errors.  Now,
in theory, this new scheme should be doing a better job at memory
management, but we may be allocating slightly more memory than before.
Before, there was a lot of editing memory in place, which is allowed,
but it means directly accessing the memory.  The new wrappers are
supposed to be memory safe, which currently disallows these direct
accesses.  I really don't think it'll matter to performance and it's
better than crashing or leaking memory.