Added error unit tests and update documentation
I added some unit tests to verify our new error handling scheme.  In
addition, I updated our documentation to discuss how to properly catch
these new errors.  Outside of that, there's some odds and ends changes
that were necessary to a variety of files in order to insure that the
manual mostly looks reasonable.  During all of this, a lot of the files
pulled in for the documentation moved, so some updates were required.