Fixed MATLAB crash and Windows MATLAB unit tests
I fixed a MATLAB crash problem that I erroneously thought was a problem
with MATLAB.  Basically, I was passing a zero length array to
mxCreateString, which isn't valid C++.  GCC doesn't care and compiles
anyway, which didn't cause any problems on Linux, but it did cause a
crash on Windows.  Strangely, the crash went away when optimizations
were turned on.  Anyway, Mathworks got me sorted, so thanks to them.

Outside of that, part of the reason why I didn't catch this problem for
some time is that I had unit tests disabled on Windows for MATLAB.
Basically, the problem is that the environment variable MATLABPATH is
treated differently on Windows than on Linux.  On Linux, we can quickly
and easily add more directories to the path.  On Windows, this sets the
path only for the location of matlabrc and only when running things from
the console.  Well, technically, it may set more paths than just
matlabrc, but matlabrc seems to blow out the other directories.  Anyway,
the way to work around this is just to prepend a bunch of addpath
commands prior to the unit tests.  In addition, there's a flag called
-wait that's needed on Windows in order to stop the MATLAB executable
from spawning a different process for MATLAB and returning immediately.
That messes up CMake, but telling the executable to wait means that
things work properly.  In any case, the whole build/unit test framework
is kind of ugly at this point, but at least it's mostly working.
Hopefully, this'll help catch more bugs in the future.