Updates to allow MacOS compilation
This update contains a series of changes to make compiling on MacOS a
bit easier.  Really, the crux of the problem is that clang doesn't do as
good of a job as gcc at determining when alias templates are equivalent.
Often, we used alias templates to either make a piece of code shorter or
to modify it slightly to make it look nicer for the manual.  Anyway, the
changes weren't a huge deal, but there were a fair number of them.
Also, clang kept throwing some warnings that I cleaned up as well.