Added MacOS and Linux installers
I added installers for both MacOS and Linux.

For Macs, this involved some tricks with messing with the rpath, but
that seems to work fine on Linux as well.  In addition, Macs are a
little funny with how environment variables are set.  It appears as
though how they're set both globally and locally has changed from
version to version.  In addition, what works on the desktop does not
work over an ssh session.  I've added some instructions on how to deal
with this to a readme file that should be immediately visible to the
user when using the MacOS installer.  Most of the time, it should just
involve copying a single plist file to one of two directories, so
hopefully that's straightforward enough.

On the Linux side, I added rpm and deb packages.  I'm not going to claim
that these are immediately ready to be uploaded to something like the
Debian package collection since these archives tend to have somewhat
strict requirements on how the packages are bundled.  That said, they
should work in a pinch and it's something that we can improve upon
later.