Fixed libraries for the MacOS installer
I fixed some library errors for the MacOS installer.  There's likely a
better way to do this, but the libraries embedded in the installer have
hard coded paths to different development libraries.  Likely, on the
deployed computer, these libraries aren't there.  As such, we have to
bundle these extra libraries and ship them with the library.  In
addition to this, we have to modify all of the libraries to point to
their new location.  I couldn't figure out a way to make CMake do this
gracefully, so I use install_name_tool to force these changes in a short
script.  This doesn't appear to be all that clean and there may be
issues if the user tries to install Optizelle somewhere other than
Applications, but it appears to work on my test system, so we'll work
with it for now.