Reworked third-party library handling
I reworked how we handle thirty-party libraries.  Before, if we built
the third-party libraries, we installed them into lib along with
liboptizelle.  On MacOS and Windows, this wasn't really a problem since
they don't follow the POSIX file system setup and have Optizelle
confined to its own library.  However, on Linux, there's a chance that
this would blow out system libraries.  Now, certainly, we could build
Optizelle to rely only on system packages and not do this, but it's a
little tricky right now for a variety of reasons.  First, we use Ubuntu
as our test system and they've gone to gcc5, which breaks ABI
compatibility with gcc4.  Since we have a MATLAB interface and that just
last month got upgraded to gcc4.9, we have a hard time hooking into the
system libraries.  If we drop MATLAB in the precompiled library, this
isn't a problem, but many customers use MATLAB, so I want MATLAB in
there.  That means that we need to compile out all of the third-party
libraries and make sure that if someone uses the rpm or deb that we
don't overwrite the system libraries.  In order to fix this, we now
install all of the libraries into a thirdparty directory inside of
share.  This way, they're self contained and won't overwrite anything.
This necessitated changes to the MacOS and Windows setups to make sure
that everything is found in the correct place.  That said, I think the
installers are working on all our supported platforms, Linux, Windows,
and MacOS, now.