Updated documentation.  Simplified license docs.
I updated the documentation to include instructions on how to use our
new installers.  In addition, I modified a fair number of the headers
and source files to strip the license information directly from the
source.  Note, the license still very much applies, but it's best if we
keep that information in a single place, which is in the licenses
folder.  That way, it's much easier for us to keep updating things like
the copyright year.  Finally, I bumped the version number and updated
the readme as we should be about ready to push to master.