Updated code current versions of dependencies
I updated the code to use the current versions of LAPACK and jsoncpp.
This meant adjusting the CMake build scripts for both.  It also meant
adjusting how we interact with the readers and writers in jsoncpp.  In
addition, I discovered a bug in how FindOctave was written where it did
not throw a fatal error for missing libraries.  I also fixed a bug with
the advanced api example, which was related to how the vectors are now
read in.