Fixed compatibility with Python 3.8 on Windows
Fixed a bug that broke compatibility with Python 3.8 on Windows.
Starting with Python 3.8, the PYTHONPATH and PATH variables are ignored
for dependencies for pyd files.  To fix this, the dependencies must
either all be in the same directory as the pyd files or we must load all
of the dependencies by hand in the correct order during initialization.
We opted for the latter approach.  If the library is compiled using a
different toolchain than mingw, the main __init__.py for the Optizelle
module will almost certainly need to be modified.